du: invert communication between DU manager and F1
The MAC notifies the DU manager about the arrival of an CCCH message
The DU manager initiates the UE creation in all DU layers
The DU manager tells the MAC to forward its buffered CCCH to the upper
layers.
The F1 now receives requests from the DU manager to start F1 setup.