PHY: Improve test vector organization
The way test vectors were organized was not uniform throughout the
tests. Now, all tests have a single .tar.gz file in the same source
directory and the test vectors are extracted in a subdirectory of the
build path.

The purpose is to automate copying test vectors from MATLAB to SRSGNB.