PHY: Extend LLR functionalities
Add default constructor, addition assignment operator and multiplication
operator by any arithmetic type.
Add the constexpr specifier to all methods that accept it.
Add free function for the dot product between a span of LLRs and a span
of an arithmetic type.
Add free function for computing the squared norm of a span of LLRs.