rlc_tx_um: pull_pdu shall not trigger buffer_state_update
This way a of large number of small SDUs in RLC queue do not
cause a storm of notifications towards MAC/scheduler. It is
the responsibility of the MAC to ask for a buffer_state_update
when required (e.g. after pulling a couple of PDUs from RLC).

Any other changes of the RLC buffer state, e.g. new SDU, ReTx,...
are still signalled to the MAC.

Related commit for RLC AM: 574daadd8296c51a06a8aad20206af74a67040ed