du,f1ap_helpers: use band_helper to set n_ta_offset value in SIB1
this fixes a bug in which some newer generation basebands have not accepted
the SIB config because it was always using the value n0 as n_ta_offset.

this value, however, is not a valid choice for TDD bands according
to 38.133 Table 7.1.2-2, resulting in some phones (with basebands
checking this value) not being able to attach in TDD mode.