f1u_local_connector: fix bug in f1u local connector that would abort the removal of f1u mapping
the bug was created because the code was checking and relying on the existance of the bearer in
the DU map. However, with the implementation of the F1AP UE context modifcation the mapping for
the DU was already removed.

the patch fixes this case by just logging the incident and continueing to remove the CU mapping anyway