pcap: fixed filename going out-of-scope
Fixed an issue where a pointer that chould go out of scope was being
captured by value in the open() function of both the MAC pcap and the
generic DLT pcap.

Now we use std::string in the open function an capture by value (copying),
avoiding the filename going out-of-scope before the open is actually
ran.