pcap: not rely open/close state on internal writer
This solves that close() could skip the stop/join procedure of
the worker if the internal writer has not yet completed the open
function, i.e. writer.is_write_enabled() results false.

If this occurs during a close() as part of the destruction, the
internal writer can be deleted while the worker is still working
with the object.