rlc_tx_am: defer deletion of released byte_buffers to ue_executor.
Upon reception of a status PDU, large amounts of byte_buffers are
removed from the tx_window. The deletion of byte_buffers involves
a certain latency to return the segments to the memory pool, which
may slow down the pcell_executor significantly.
This commit postpones this time-consuming destruction to the ue_executor.