prach_detector: modify results structure
We used to have power and SINR fields, which were not estimated
properly. We replace them with a more meaningful detection_metric field,
which is the ratio between the detection metric and the threshold - in
other words, the more the value is larger than one, the more reliable
the detection is.