gnb,sctp: expose option to set/unset SCTP_NODELAY
With NODELAY enabled, SCTP messages are sent as soon as possible with no
unnecessary delay, at the cost of transmitting more packets over the
network. Otherwise their transmission might be delayed and concatenated
with subsequent messages in order to transmit them in one big PDU.

If the local interface supports jumbo frames (MTU size > 1500) but not
the receiver, then the receiver might discard big PDUs and the stream
might get stuck.