rlc,nru: notify "desired buffer size for DRB" according to TS 38.425
This value informs the CU of the amount of data per DRB to be sent beyond
the highest delivered PDCP SN (AM) or highest transmitted PDCP SN (UM).
It corresponds to the total capacity of the RLC queue (in bytes).

TS 38.245 Sec. 5.4.2.1:
(...)
- If the value of the desired buffer size is 0, the hosting node shall stop sending any data per bearer.
- If the value of the desired buffer size in b) above is greater than 0, the hosting node may send up to this
amount of data per bearer beyond the "Highest Delivered NR PDCP SN" for RLC AM, or the hosting node
may send up to this amount of data per bearer beyond the "Highest Transmitted NR PDCP SN" for RLM
UM.
(...)